Move Config to utils
